Convert all code using handlers to use Erebot_Callable (#381)
All code that uses Erebot_(Event|Raw)_Handler now expects the $callback argument
to be given as an Erebot_Callable instance.
This revision also fixes an issue in the WatchList module when the ISON strategy
is used and the server returns an empty result.
Last but not least, it also fixes a race condition in the bot when disconnecting
(the race condition was introduced in r1013 while removing the core's dependency
on the IrcConnector module).