Commands to change channel privileges.
Adds support for the following commands:

* `!op`
* `!deop`
* `!halfop`
* `!dehalfop`
* `!voice`
* `!devoice`
* `!protect`
* `!deprotect`
* `!owner`
* `!deowner`

All of these are usually provided by IRC services such as Anope.
For now, the bot does not check whether it can actually change those modes
before sending the command to the server (adding such a check would be
troublesome with little gain: the command results in a no-op if the bot
has insufficient privileges).
Fixes #1.